# openapi.yaml
openapi: '3.1.0'
info:
  title: 'Mission Control API'
  version: '1.0.0'
  description: |-
    Generated from the Backend’s Fastify route table by `pnpm api:spec` (F5.1). Do not edit
    by hand: `pnpm api:spec:check` fails when this file and the routes disagree, and
    `pnpm api:types:check` fails when apps/frontend/src/lib/api/types.generated.ts does.

    REQUEST shapes are the schemas Ajv actually enforces, so they are exact.

    RESPONSE shapes are the schemas each route declares under `schema.response`. Read the
    guarantee precisely: they are DECLARATIONS, NOT SERIALIZERS. The Backend installs a
    serializer compiler that ignores them and calls JSON.stringify, so a schema here can never
    remove a field the handler produced — the failure mode a Fastify response schema normally
    carries. What keeps them true instead is a conformance check that validates every reply
    against its own schema in the test tiers, in both directions: an undeclared property and a
    missing required property are both failures. See apps/backend/src/http/response-schema.ts.

    Operations still marked `x-mc-response-schema: undeclared` are the ones with no declaration,
    and the marker exists so that gap stays countable. Today they are: GET /health,
    GET /services/health and GET /schedule (owned by modules outside this change),
    GET|PUT /settings/{category} (the GET serves one of five different documents, which needs a
...
    and GET /ws.

...
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /agent-teams:
    get:
      operationId: getAgentTeams
      summary: 'GET /agent-teams'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum:
              - global
              - project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentTeam'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postAgentTeams
      summary: 'POST /agent-teams'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - scope
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 2000
                scope:
                  type: string
                  enum:
                    - global
                    - project
                projectId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                agentIds:
                  type: array
                  maxItems: 50
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                projectIds:
                  type: array
                  maxItems: 100
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentTeam'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/agent-teams/{id}':
    get:
      operationId: getAgentTeamsById
      summary: 'GET /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentTeam'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    patch:
      operationId: patchAgentTeamsById
      summary: 'PATCH /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 2000
                agentIds:
                  type: array
                  maxItems: 50
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                projectIds:
                  type: array
                  maxItems: 100
                  items:
                    type: string
                    pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/AgentTeam'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    delete:
      operationId: deleteAgentTeamsById
      summary: 'DELETE /agent-teams/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '204':
          description: 'No Content'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  /agents:
    get:
      operationId: getAgents
      summary: 'GET /agents'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum:
              - global
              - project
              - session
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postAgents
      summary: 'POST /agents'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - scope
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 2000
                scope:
                  type: string
                  enum:
                    - global
                    - project
                    - session
                projectId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                sessionId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                runtime:
                  type: string
                  enum:
                    - claude_code
                permissions:
                  type: object
                  additionalProperties: false
                  properties:
                    repository:
                      type: object
                      additionalProperties: false
                      properties:
                        read:
                          type: boolean
                        write:
                          type: boolean
                        shell:
                          type: boolean
                instructions:
                  type:
                    - string
                    - 'null'
                  maxLength: 20000
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  '/projects/{id}':
    get:
      operationId: getProjectsById
      summary: 'GET /projects/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    patch:
      operationId: patchProjectsById
      summary: 'PATCH /projects/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                description:
                  type:
                    - string
                    - 'null'
                  maxLength: 10000
                workflowMode:
                  type:
                    - string
                    - 'null'
                  enum:
                    - manual
                    - assisted
                    - null
                archivedAt:
                  type:
                    - string
                    - 'null'
                  maxLength: 64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    delete:
      operationId: deleteProjectsById
      summary: 'DELETE /projects/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '204':
          description: 'No Content'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/projects/{id}/available-agents':
    get:
      operationId: getProjectsByIdAvailableAgents
      summary: 'GET /projects/{id}/available-agents'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/ProjectAvailableAgents'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  /pull-requests:
    get:
      operationId: getPullRequests
      summary: 'GET /pull-requests'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: repositoryId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - open
              - merged
              - closed
              - draft
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PullRequest'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
components:
  schemas:
...
    Agent:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - scope
        - projectId
        - sessionId
        - runtime
        - permissions
        - disallowedTools
        - instructions
        - archivedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        scope:
          $ref: '#/components/schemas/AgentScope'
        projectId:
          type:
            - string
            - 'null'
        sessionId:
          type:
            - string
            - 'null'
        runtime:
          $ref: '#/components/schemas/AgentRuntime'
        permissions:
          $ref: '#/components/schemas/AgentPermissions'
        disallowedTools:
          type: array
          items:
            type: string
          description: 'Derived from `permissions` by the same function the launch path calls, and never accepted on a write. `permissions` says what was asked for; this says what actually happens to the runtime, so the model can be audited rather than trusted.'
        instructions:
          type:
            - string
            - 'null'
        archivedAt:
          type:
            - string
            - 'null'
        createdAt:
          type: string
        updatedAt:
          type: string
...
    AgentTeam:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - scope
        - projectId
        - members
        - projectIds
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        scope:
          $ref: '#/components/schemas/AgentTeamScope'
        projectId:
          type:
            - string
            - 'null'
        members:
          type: array
          items:
            $ref: '#/components/schemas/AgentTeamMember'
        projectIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    AgentTeamMember:
      type: object
      additionalProperties: false
      required:
        - agentId
        - name
        - scope
        - projectId
        - runtime
        - archivedAt
        - addedAt
      properties:
        agentId:
          type: string
        name:
          type: string
        scope:
          $ref: '#/components/schemas/AgentScope'
        projectId:
          type:
            - string
            - 'null'
        runtime:
          $ref: '#/components/schemas/AgentRuntime'
        archivedAt:
          type:
            - string
            - 'null'
        addedAt:
          type: string
    AgentTeamScope:
      type: string
      enum:
        - global
        - project
...
    ListEnvelopeMeta:
      type: object
      additionalProperties: false
      required:
        - nextCursor
        - limit
      properties:
        nextCursor:
          type:
            - string
            - 'null'
        limit:
          type: integer
...
    Project:
      type: object
      additionalProperties: false
      required:
        - id
        - workspaceId
        - name
        - description
        - workflowMode
        - createdAt
        - updatedAt
        - archivedAt
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        workflowMode:
          anyOf:
            - $ref: '#/components/schemas/WorkflowMode'
            - type: 'null'
          description: 'null means INHERIT integrations.github.workflowMode (arbitration A10), not "no mode". Three-valued by design: a two-valued field could not express "follow the global default".'
        createdAt:
          type: string
        updatedAt:
          type: string
        archivedAt:
          type:
            - string
            - 'null'
    ProjectAvailableAgents:
      type: object
      additionalProperties: false
      required:
        - projectId
        - team
        - agents
      properties:
        projectId:
          type: string
        team:
          anyOf:
            - $ref: '#/components/schemas/AssignedTeam'
            - type: 'null'
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AvailableAgent'
    PullRequest:
      type: object
      additionalProperties: false
      required:
        - id
        - repositoryId
        - number
        - title
        - url
        - state
        - authorLogin
        - sourceBranch
        - targetBranch
        - openedAt
        - reviewedAt
        - mergedAt
        - closedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        repositoryId:
          type: string
        number:
          type: integer
        title:
          type: string
        url:
          type:
            - string
            - 'null'
        state:
          $ref: '#/components/schemas/PullRequestState'
        authorLogin:
          type:
            - string
            - 'null'
        sourceBranch:
          type:
            - string
            - 'null'
        targetBranch:
          type:
            - string
            - 'null'
        openedAt:
          type:
            - string
            - 'null'
        reviewedAt:
          type:
            - string
            - 'null'
        mergedAt:
          type:
            - string
            - 'null'
        closedAt:
          type:
            - string
            - 'null'
        createdAt:
          type: string
        updatedAt:
          type: string
...
  responses:
    ErrorEnvelope:
      description: 'Any non-2xx response (F5.4). The status code and the `error.code` registry entry carry the meaning; see docs/tds/04-api-contracts-and-events.md §1.3.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mc_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'mct_ API token'